requirements.txt
beautifulsoup4==4.12.3
chromadb==0.4.24
Django==5.0.3
djangorestframework==3.14.0
langdetect==1.0.9
nltk==3.8.1
numpy==1.26.4
openai==1.25.1
pdfplumber==0.11.0
python-dotenv==1.0.1
Requests>=2.32.0
soundfile==0.12.1
textstat==0.7.3
transformers==4.38.2
torch~=2.2.1
fasttext~=0.9.2